cast

+++ Userland/SampleCodeModule/libSysCalls.c
@@ -5,13 +5,13 @@ extern uint64_t sysCall(uint64_t rax, uint64_t rdi, uint64_t rsi , uint64_t rdx 
 // deberia usar el strlen en vez del 1 despues del buffer.
 // aca en vez de 1 deberia ser 4
 void print(char* buffer){
-    sysCall(1,1,buffer,1,0,0); 
+    sysCall(1,1,(uint64_t)buffer,1,0,0); 
    
 }
 
 //0 en rax, en realidad deberia ser 3
 uint64_t readBuffer(char* buffer, uint64_t length){
-    return sysCall(0,0,buffer,0,length,0);  //nose que hace read la verdad
+    return sysCall(0,0,(uint64_t)buffer,0,length,0);  //nose que hace read la verdad
 }
 
 void clearScreen(){
